// src/lib/api/types/tagResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Adisu Backend API
 * Backend API for the Adisu platform
 * OpenAPI spec version: 1.0.0
 */

export interface TagResponse {
  /** A URL to the JSON Schema for this object. */
  readonly $schema?: string;
  /** Created timestamp */
  createdAt?: string;
  /** Amharic description */
  descAm?: string;
  /** English description */
  descEn?: string;
  /** Tag group */
  group: string;
  /** Icon identifier */
  icon?: string;
  /** Tag ID */
  id: string;
  /** Whether tag is active */
  isActive: boolean;
  /** Whether multiple tags from this group can be selected together */
  isMultiSelect: boolean;
  /** Amharic name */
  nameAm: string;
  /** English name */
  nameEn: string;
  /** Tag slug */
  slug: string;
  /** Display order */
  sortOrder: number;
  /** Updated timestamp */
  updatedAt?: string;
}